Move chip wiring to where it belongs + document new types
See #16
